Added an example of hello world without BIOS calls
Example uses the fact that BIOS maps video memory to address 0xB8000. We
can then move 0xB800 to a segment register and use segment:offset
addressing to access this memory. Then we can show characters by
treating 0xB800:0000 as a uint16_t array, where low 8 bytes is the ASCII
character, and the high 8 bytes is the color attribute of this
character.